config/urls.py
from django.conf.urls import include, url
from django.contrib import admin
from rest_framework_jwt.views import obtain_jwt_token, refresh_jwt_token, verify_jwt_token

from config.router import router
from users.apis import register_by_access_token

urlpatterns = [
    url(r'^admin/', include(admin.site.urls)),
    url(r"^", include(router.urls)),
    # url(r'^social/', include('social.apps.django_app.urls', namespace='social')),
    url(r'^api/login/', include('rest_social_auth.urls_jwt')),
    url(r'^api-token-auth/', obtain_jwt_token),
    url(r'^api-token-refresh/', refresh_jwt_token),
    url(r'^api-token-verify/', verify_jwt_token),
]
