requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
defusedxml==0.4.1
Django==1.9.6
django-cors-headers==1.1.0
django-jet==0.1.4
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
kombu==3.0.35
mysqlclient==1.3.7
oauthlib==1.1.1
PyJWT==1.4.0
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
qiniu==7.0.7
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
rest-social-auth==0.4.4
shortuuid==0.4.3
six==1.10.0
uWSGI==2.0.13.1
